# backend/core/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma de cinema screeK.",
        "title": "screeK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte screeK",
            "url": "http://www.screek.com/support",
            "email": "suporte@screek.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
...
        "/movies/discover": {
            "get": {
                "description": "Filtra filmes por gênero e/ou ano de lançamento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Descoberta de filmes com filtros",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Gênero (TMDB ID)",
                        "name": "genre_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ano de Lançamento",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/movies.Movie"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/search": {
            "get": {
                "description": "Realiza uma busca textual no catálogo de filmes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Busca filmes por título",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Termo de busca (ex: Batman)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "MOVIE",
                        "description": "Tipo de busca (MOVIE, PERSON, USER, LIST)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/movies.Movie"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Retorna informações do TMDB e cache local para um filme específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Detalhes completos de um filme",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Filme (TMDB ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movies.Movie"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/movies/{id}/recommendations": {
            "get": {
                "description": "Retorna uma lista de filmes recomendados baseados em um filme específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Recomendações de filmes similares",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Filme",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/movies.TMDBMovie"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna as últimas 20 notificações (lidas e não lidas).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Lista notificações do usuário",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/notifications.Notification"
                            }
                        }
                    }
                }
            }
        },
        "/notifications/read-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Marca TODAS como lidas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}/read": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Marca uma notificação como lida",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Notificação",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/people/{id}": {
            "get": {
                "description": "Retorna biografia, foto e informações do TMDB via cache local.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Busca detalhes de uma pessoa (ator/diretor)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Pessoa (TMDB ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movies.Person"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/people/{id}/movies": {
            "get": {
                "description": "Retorna a filmografia (créditos) de um ator ou membro da equipe.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Lista filmes relacionados a uma pessoa",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Pessoa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/movies.TMDBPersonMovieCast"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/movies.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ws": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Realiza o upgrade HTTP para WebSocket para receber alertas em tempo real.",
                "tags": [
                    "Notifications"
                ],
                "summary": "Conexão WebSocket para Notificações",
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "httputil.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "httputil.Role": {
            "type": "string",
            "enum": [
                "USER",
                "MANAGER",
                "ADMIN"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleManager",
                "RoleAdmin"
            ]
        },
...
        "movies.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "movies.Genre": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tmdb_id": {
                    "type": "integer"
                }
            }
        },
        "movies.Movie": {
            "type": "object",
            "properties": {
                "credits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movies.MovieCredit"
                    }
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movies.Genre"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "original_language": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_url": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "runtime": {
                    "type": "integer"
                },
                "spoken_languages": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tmdb_id": {
                    "type": "integer"
                }
            }
        },
        "movies.MovieCredit": {
            "type": "object",
            "properties": {
                "character": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movie_id": {
                    "type": "integer"
                },
                "person": {
                    "$ref": "#/definitions/movies.Person"
                },
                "person_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "movies.Person": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "profile_url": {
                    "type": "string"
                },
                "tmdb_id": {
                    "type": "integer"
                }
            }
        },
        "movies.TMDBMovie": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "original_language": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vote_average": {
                    "type": "number"
                }
            }
        },
        "movies.TMDBPersonMovieCast": {
            "type": "object",
            "properties": {
                "character": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "original_title": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vote_average": {
                    "type": "number"
                }
            }
        },
        "notifications.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_read": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/users.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "users.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "default_city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "favorite_movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movies.Movie"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "pronouns": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/httputil.Role"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}